requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.1.0
pydantic_core==2.18.2
python-dotenv==1.0.0
PyYAML==6.0.1
setuptools==69.2.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.43.0
